+++ rust/template/ddlog_derive/tests/ui/fail/duplicated_attrs.rs
@@ -1,23 +1,24 @@
 use ddlog_derive::{FromRecord, IntoRecord};
+use serde::Deserialize;
 
 fn main() {}
 
-#[derive(FromRecord, IntoRecord)]
+#[derive(FromRecord, IntoRecord, Deserialize)]
 #[ddlog(rename = "Foo")]
 #[ddlog(rename = "Foo")]
 struct Foo {}
 
-#[derive(IntoRecord)]
+#[derive(IntoRecord, Deserialize)]
 #[ddlog(into_record = "Bar")]
 #[ddlog(rename = "Foo")]
 struct Bar {}
 
-#[derive(IntoRecord)]
+#[derive(IntoRecord, Deserialize)]
 #[ddlog(from_record = "Bar")]
 #[ddlog(rename = "Foo")]
 struct Baz {}
 
-#[derive(FromRecord)]
+#[derive(FromRecord, Deserialize)]
 #[ddlog(rename = "Foo")]
 #[ddlog(into_record = "Biz")]
 struct Biz {}